Trait pallet_xdns_rpc::XdnsApiServer

pub trait XdnsApiServer<AccountId>: Sized + Send + Sync + 'static {
    fn fetch_records(&self) -> RpcResult<FetchXdnsRecordsResponse<AccountId>>;
    fn fetch_abi(&self, chain_id: ChainId) -> RpcResult<GatewayABIConfig>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        AccountId: Send + Sync + 'static + Serialize,
    { ... }
}

Server trait implementation for the XdnsApi RPC API.

Required Methods

fn fetch_records(&self) -> RpcResult<FetchXdnsRecordsResponse<AccountId>>

Returns all known XDNS records

fn fetch_abi(&self, chain_id: ChainId) -> RpcResult<GatewayABIConfig>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
AccountId: Send + Sync + 'static + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.